+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Remove `enable_set_attribute_value` config option
 - Implement `C_SetAttributeValue` for `CKA_ID` to support renaming keys
 - Remove corresponding certificate and public key objects from the cache if a private key is deleted ([#260](https://github.com/Nitrokey/nethsm-pkcs11/issues/260))
+- Add `CKF_ENCRYPT` flag for `CKM_RSA_PKCS`
 
 ## [1.7.2][] (2025-07-25)